+++ package.json
@@ -59,11 +59,8 @@
 		]
 	},
 	"devDependencies": {
-		"eslint-config-prettier": "^8.1.0",
-		"eslint-plugin-prettier": "^3.3.1",
 		"husky": "^5.1.3",
 		"lint-staged": "^10.5.4",
-		"prettier": "^2.2.1",
 		"pretty-quick": "^3.1.0"
 	},
 	"husky": {
