+++ src/components/ResizableComponent/resizable.tsx
