+++ src/index.tsx
@@ -2,7 +2,7 @@ import ReactDOM from "react-dom";
 import { Provider } from 'react-redux'
 import { store } from './core'
 // import CodeVaultComponent from "./components/codevault";
-import MarkDownEditor from './components/texteditor'
+import MarkDownEditor from './components/TextEditorComponent/texteditor'
 import "bulmaswatch/darkly/bulmaswatch.min.css";
